Move components and systems outside ECS.
ECS should never directly invoke systems, because ECS’s responsibility is just managing entities and components. It doesn't know or care what logic is applied.
