+++ internal/game/game.go
@@ -7,9 +7,9 @@ import (
 
 	"github.com/juanancid/maze-adventure/internal/config"
 	"github.com/juanancid/maze-adventure/internal/ecs"
-	"github.com/juanancid/maze-adventure/internal/ecs/systems"
 	"github.com/juanancid/maze-adventure/internal/events"
 	"github.com/juanancid/maze-adventure/internal/levels"
+	"github.com/juanancid/maze-adventure/internal/systems"
 )
 
 type Game struct {
