+++ internal/levels/factory.go
@@ -3,8 +3,8 @@ package levels
 import (
 	"github.com/hajimehoshi/ebiten/v2"
 
+	"github.com/juanancid/maze-adventure/internal/components"
 	"github.com/juanancid/maze-adventure/internal/ecs"
-	"github.com/juanancid/maze-adventure/internal/ecs/components"
 	"github.com/juanancid/maze-adventure/internal/layout"
 	"github.com/juanancid/maze-adventure/internal/utils"
 )
