+++ internal/systems/maze_renderer.go
